requirements.txt
Django~=3.1
djangorestframework~=3.11.1
requests~=2.24.0
beautifulsoup4~=4.9.1
html5lib~=1.1
psycopg2~=2.8.5
gunicorn~=20.0.4
django-cors-headers~=3.4.0